// packages/types-generator/src/analyzers/sqlite/types.ts
export interface TableInfo {
  name: string
  columns: ColumnInfo[]
  relations: RelationInfo[]
}

export interface ColumnInfo {
  name: string
  type: string
  notNull: boolean
}

export interface RelationInfo {
  fieldName: string
  targetTable: string
  type: 'one-to-one' | 'one-to-many'
}